Refactor code structure for improved readability and maintainability

+++ src/Modules/Download-Keys/Service/delete-download-keys.service.ts
@@ -1,12 +1,13 @@
 // get-download-key.service.ts
 import { Injectable, BadRequestException, ConflictException, HttpException, InternalServerErrorException, Inject, NotFoundException } from "@nestjs/common";
 import { IDownloadKeysRepositories } from "../Repositories/IDownload-keys-repositories";
+import { Role } from "src/Modules/Auth/Guards/roles.enum";
 
 @Injectable()
 export class DeleteDownloadKeyService {
   constructor(private readonly repository: IDownloadKeysRepositories) {}
 
-  async execute(params: Partial<{ id: string; usedByIp: string; key: string }>): Promise<any>
+  async execute(params: Partial<{ id: string; usedByIp: string; key: string }>, credentials?:{sub: string, role: Role}): Promise<any>
   {
     try
     {   
@@ -14,6 +15,11 @@ export class DeleteDownloadKeyService {
         if (!hasParam)
         {
             throw new BadRequestException("Forneça pelo menos um parâmetro de busca: id, usedByIp ou key.");
+        }
+        const existsKey = await this.repository.getDownloadKey(params)
+        if(!existsKey)
+        {
+            throw new NotFoundException("Esta chave de download não existe.")
         }
         const result = await this.repository.deleteDownloadKey(params);
         if (!result){throw new NotFoundException("Chave de download não encontrada.");}
@@ -24,6 +30,7 @@ export class DeleteDownloadKeyService {
             {
                 throw error
             }
+            console.log(error)
             throw new InternalServerErrorException("Ocorreu um erro, tente novamente")
     }
   }
